+++ packages/runtime-v8-isolated/tests/fetchevent.ts
@@ -140,4 +140,88 @@ test('FetchEvent', async (t) => {
             returnValue: undefined,
         });
     });
+
+    await t.test('should accept a Uint8Array body (as json)', async () => {
+        const code = `
+            addEventListener('fetch', async (event) => {
+                const payload = await event.request.json();
+                return new Response('Hello ' + payload.name);
+            });
+            `;
+
+        const body = new TextEncoder().encode(JSON.stringify({ name: 'John' }));
+
+        const result = await runIsolatedEvent(code, {
+            type: 'fetch',
+            request: {
+                body,
+                headers: {
+                    'Content-Type': 'application/json',
+                },
+            },
+        });
+
+        assert.deepEqual(result, {
+            logs: [],
+            returnValue: {
+                body: 'Hello John',
+                headers: {},
+                status: 200,
+            },
+        });
+    });
+
+    await t.test('should accept a Uint8Array body', async () => {
+        const code = `
+            addEventListener('fetch', async (event) => {
+                const buf = await event.request.arrayBuffer();
+                return new Response('Hello ' + buf.byteLength);
+            });
+            `;
+
+        const body = new Uint8Array([0, 1, 2, 3, 4, 5, 6, 7, 8, 9]);
+
+        const result = await runIsolatedEvent(code, {
+            type: 'fetch',
+            request: {
+                body,
+            },
+        });
+
+        assert.deepEqual(result, {
+            logs: [],
+            returnValue: {
+                body: 'Hello 10',
+                headers: {},
+                status: 200,
+            },
+        });
+    });
+
+    await t.test('should accept a Uint8Array body (as text)', async () => {
+        const code = `
+            addEventListener('fetch', async (event) => {
+                const text = await event.request.text();
+                return new Response('Hello ' + text);
+            });
+            `;
+
+        const body = new TextEncoder().encode('John Doe');
+
+        const result = await runIsolatedEvent(code, {
+            type: 'fetch',
+            request: {
+                body,
+            },
+        });
+
+        assert.deepEqual(result, {
+            logs: [],
+            returnValue: {
+                body: 'Hello John Doe',
+                headers: {},
+                status: 200,
+            },
+        });
+    });
 });