+++ packages/runtime-v8-stdlib/src/body_mixin.ts
@@ -75,7 +75,7 @@ export class BodyMixin implements Body {
             return this.bodySource;
         }
 
-        const arr = await this.arrayBuffer();
+        const arr = new Uint8Array(await this.arrayBuffer());
         return new TextDecoder('utf-8').decode(arr);
     }
 
@@ -102,12 +102,14 @@ export class BodyMixin implements Body {
         } else if (typeof this.bodySource === 'string') {
             const enc = new TextEncoder();
             return enc.encode(this.bodySource).buffer as ArrayBuffer;
-        } else if (this.bodySource instanceof ReadableStream) {
-            return bufferFromStream(this.bodySource.getReader());
-        } else if (this.bodySource instanceof FormData) {
-            const enc = new TextEncoder();
-            return enc.encode(this.bodySource.toString()).buffer as ArrayBuffer;
-        } else if (!this.bodySource) {
+        }
+        // else if (this.bodySource instanceof ReadableStream) {
+        //     return bufferFromStream(this.bodySource.getReader());
+        // } else if (this.bodySource instanceof FormData) {
+        //     const enc = new TextEncoder();
+        //     return enc.encode(this.bodySource.toString()).buffer as ArrayBuffer;
+        // }
+        else if (!this.bodySource) {
             return new ArrayBuffer(0);
         }
         throw new Error(`Body type not yet implemented: ${this.bodySource.constructor.name}`);
@@ -116,23 +118,21 @@ export class BodyMixin implements Body {
 
 /** @hidden */
 function validateBodyType(owner: any, bodySource: any) {
-    // if (
-    //     bodySource instanceof Int8Array ||
-    //     bodySource instanceof Int16Array ||
-    //     bodySource instanceof Int32Array ||
-    //     bodySource instanceof Uint8Array ||
-    //     bodySource instanceof Uint16Array ||
-    //     bodySource instanceof Uint32Array ||
-    //     bodySource instanceof Uint8ClampedArray ||
-    //     bodySource instanceof Float32Array ||
-    //     bodySource instanceof Float64Array
-    // ) {
-    //     return true;
-    // } else if (bodySource instanceof ArrayBuffer) {
-    //     return true;
-    // } else
-
-    if (typeof bodySource === 'string') {
+    if (
+        bodySource instanceof Int8Array ||
+        bodySource instanceof Int16Array ||
+        bodySource instanceof Int32Array ||
+        bodySource instanceof Uint8Array ||
+        bodySource instanceof Uint16Array ||
+        bodySource instanceof Uint32Array ||
+        bodySource instanceof Uint8ClampedArray ||
+        bodySource instanceof Float32Array ||
+        bodySource instanceof Float64Array
+    ) {
+        return true;
+    } else if (bodySource instanceof ArrayBuffer) {
+        return true;
+    } else if (typeof bodySource === 'string') {
         return true;
     }
     // else if (bodySource instanceof ReadableStream) {
@@ -146,38 +146,38 @@ function validateBodyType(owner: any, bodySource: any) {
     throw new Error(`Bad ${owner.constructor.name} body type: ${bodySource.constructor.name}`);
 }
 
-function bufferFromStream(stream: ReadableStreamReader): Promise<ArrayBuffer> {
-    return new Promise((resolve, reject) => {
-        const parts: Uint8Array[] = [];
-        const encoder = new TextEncoder();
-        // recurse
-        (function pump() {
-            stream
-                .read()
-                .then(({ done, value }) => {
-                    if (done) {
-                        return resolve(concatenate(...parts));
-                    }
+// function bufferFromStream(stream: ReadableStreamReader): Promise<ArrayBuffer> {
+//     return new Promise((resolve, reject) => {
+//         const parts: Uint8Array[] = [];
+//         const encoder = new TextEncoder();
+//         // recurse
+//         (function pump() {
+//             stream
+//                 .read()
+//                 .then(({ done, value }) => {
+//                     if (done) {
+//                         return resolve(concatenate(...parts));
+//                     }
 
-                    if (typeof value === 'string') {
-                        parts.push(encoder.encode(value));
-                    } else if (value instanceof ArrayBuffer) {
-                        parts.push(new Uint8Array(value));
-                    } else if (!value) {
-                        // noop for undefined
-                    } else {
-                        console.log('unhandled type on stream read:', value);
-                        reject('unhandled type on stream read');
-                    }
+//                     if (typeof value === 'string') {
+//                         parts.push(encoder.encode(value));
+//                     } else if (value instanceof ArrayBuffer) {
+//                         parts.push(new Uint8Array(value));
+//                     } else if (!value) {
+//                         // noop for undefined
+//                     } else {
+//                         console.log('unhandled type on stream read:', value);
+//                         reject('unhandled type on stream read');
+//                     }
 
-                    return pump();
-                })
-                .catch((err) => {
-                    reject(err);
-                });
-        })();
-    });
-}
+//                     return pump();
+//                 })
+//                 .catch((err) => {
+//                     reject(err);
+//                 });
+//         })();
+//     });
+// }
 
 /** @hidden */
 function concatenate(...arrays: Uint8Array[]): ArrayBuffer {