+++ packages/runtime-v8-stdlib/src/bridge.ts
@@ -17,6 +17,15 @@ declare function __fetch(
     callback: (err: string | undefined, request: any, body: any) => void
 ): void;
 
+declare function __encodeStringToBuffer(
+    input: string
+): Uint8Array;
+
+declare function __decodeBufferToString(
+    input: Uint8Array,
+    encoding?: string
+): string;
+
 // Grab a reference to the ivm module and delete it from global scope. Now this closure is the
 // only place in the context with a reference to the module. The `ivm` module is very powerful
 // so you should not put it in the hands of untrusted code.
@@ -32,4 +41,10 @@ delete global.__addEventListener;
 const _fetch = __fetch;
 delete global.__fetch;
 
-export { _ivm, _log, _addEventListener, _fetch };
+const _encodeStringToBuffer = __encodeStringToBuffer;
+delete global.__encodeStringToBuffer;
+
+const _decodeBufferToString = __decodeBufferToString;
+delete global.__decodeBufferToString;
+
+export { _ivm, _log, _addEventListener, _fetch, _encodeStringToBuffer, _decodeBufferToString };